Clear the cached bluetooth devices due to UI bug
This is a work-around to fix the Modello Bluetooth settings UI from
hanging after bluetooth scan.  The cached bluetooth device objects that
are being serialized using the Knockout library is not cached properly,
this caused Modello to load a list of invalid devices.  This fix clears
the cached devices and starts fresh when it loads.

Change-Id: I80115e4510a504e51c4ff7a78cb640cfe517a844
Signed-off-by: Jimmy Huang <jimmy.huang@intel.com>